Add more FAQs for people using a common spelling
Added `datapack structure`, `datapack structure (light)`,
`resourcepack structure`, and `resourcepack structure (light)`.

Additionally, added `data-pack` and `data-packs` to the tags for all
variants of `data pack structure`, `resource-pack` and `resource-packs`
to the tags for all variants of `resource pack structure`, and `light`
and `(light)` to the tags for all light mode variants of
`data pack structure` and `resource pack structure`.